+++ packages/fullScreen/web/src/icons/index.js
@@ -1,5 +1,5 @@
-export { default as CloseIcon } from './close';
-export { default as ExpandIcon } from './expand';
-export { default as ShrinkIcon } from './shrink';
-export { default as ArrowLeft } from './arrowLeft';
-export { default as ArrowRight } from './arrowRight';
+export { default as CloseIcon } from './CloseIcon';
+export { default as ExpandIcon } from './ExpandIcon';
+export { default as ShrinkIcon } from './ShrinkIcon';
+export { default as ArrowLeft } from './ArrowLeft';
+export { default as ArrowRight } from './ArrowRight';